fix: object file
