+++ src/utils/index.ts
@@ -1,4 +1,4 @@
 export * from './format';
-
 export * from './maximal-paths';
+export * from './object';
 export * from './serialization';
